+++ srcs/ambient_lightning.c
@@ -1,6 +1,6 @@
 #include "../includes/minirt.h"
 
-t_a	*a_default(float ratio, float color)
+t_a	*a_default(float ratio, int color)
 {
 	t_a	*new_a;
 	new_a = (t_a *)malloc(sizeof(t_a));
