+++ srcs/sphere.c
@@ -1,6 +1,6 @@
 #include "../includes/minirt.h"
 
-t_sp 	*sp_default(t_vec *center, float diameter, float color)
+t_sp 	*sp_default(t_vec *center, float diameter, int color)
 {
 	t_sp 	*sp;
 
@@ -12,7 +12,7 @@ t_sp 	*sp_default(t_vec *center, float diameter, float color)
 	return (sp);
 }
 
-void 	sp_list(t_sp **sp_list, t_vec *center, float diameter, float color)
+void 	sp_list(t_sp **sp_list, t_vec *center, float diameter, int color)
 {
 	t_sp	*new_sp;
